+++ personnel_on_boarding/src/com/pravin/service/CityService.java
@@ -13,6 +13,6 @@ public interface CityService {
 	public List<City> getAll();
 	public City getById(long id);
 	
-	public List<City> getByStateId(long id);
+	public List<City> getByStateId(int id);
 	
 }
